Add logging module
The idea of the logging module is that instead of commenting out the
debugging output, there's a configuration file that specifies what loglevel
is desired for a given file/function.

Of course, when having a fancy config file is slow as hell.  But by the
time this tool is actually useful it should be trivial to disable the
module completely...