Refactor decode_fde/unwind_find_caller_offset
Refactored into:
 * decode_fde_boundaries
 * decode_fde_cfa
 * decode_cfa_opcodes

Right now, FDEs are decoded every time that a symbol needs to be found.  By
doing this refactoring, the decoding should be easier to do offline
(that is, before ptrace(2)ing), so that the decoded info can be used
separately.  Note that implementation for that is still missing, though.